# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Симптомчеккер API",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/index": {
            "get": {
                "description": "Первичное отображение данных на странице (разделы и подразделы)",
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/section": {
            "get": {
                "description": "Получение подразделов относящихся к выбранной части тела",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID раздела (части тела). Так же принимает значение * (звёздочка) - которое указывает на получение ВСЕХ подразделов"
                    },
                    {
                        "name": "sex",
                        "in": "query",
                        "description": "Выбор пола.Необзательный параметр принимает значения \"w\", \"m\" (если оставить пустым - то будут выведены записи и женские и мужские подразделы)"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    },
                    "404": {
                        "description": "error"
                    }
                }
            }
        },
        "/api/v1/subsection": {
            "get": {
                "description": "Получение первого вопроса относящегося к выбранному подразделу",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID подраздела"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        },
        "/api/v1/answer": {
            "get": {
                "description": "Получение диагноза на основании результатов к ответам, либо если диагноз не удалось установить - то происходит выдача следующего вопроса. Данный API провоцируется нажатием на ответы «ДА» или «НЕТ» к вопросам",
                "parameters": [
                    {
                        "name": "subsection",
                        "in": "query",
                        "description": "Введите ID подраздела"
                    },
                    {
                        "name": "questions",
                        "in": "query",
                        "description": "Отправьте массив JSON вида: {\"11\":0, \"12\":0, \"13\":1}, где первое значение элементов массива в кавычках - это ID вопроса, а второе значение (0 или 1) - ответ НЕТ или ДА на вопрос"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The data"
                    }
                }
            }
        }
    }
}